+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
 ## 5.0.0 - *2021-01-20*
 
 **Breaking changes, please see [UPGRADING.md](./UPGRADING.md).**