+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## 5.0.1 - *2021-01-20*
 
 ## 5.0.0 - *2021-01-20*